feat: Display error message if saving fails
When the save of the query ends up in error, then the error message
should be displayed and not silently ignored.

To test, you can raise a RestException at the beginning of the
`TimetrackingManagementWidgetResource::put()` method.

On save =>
* The error message is displayed
* The editor stays open

If the editor is closed, then error message is not displayed anymore.

Part of story #41084: Retrieve widget configuration

Change-Id: I55aa0754c0b3d7892861b35f832dc9cf43396fb2