fix: request #44444 Artidoc section cannot be empty at heading level change
With a tracker whose description field is not required.

When you change heading level of an artifact with an empty description
in an artidoc you cannot save it.

The artifact section description is empty string "" and not empty html
"<p></p>". But Restler consider empty string as no value and thus return
an error 400 as we require a description.

By making the description parameter non-required we can now send an
empty string

*Testing:*

- Have a tracker with a non required description field
- Have an artifact of this tracker with an empty description
- Import this artifact in an artidoc and change its heading level
- Save should work

Change-Id: If3803d107ac5c1352813f84ed8f86d1875c46d7e