feat: Configure semantic description in TestCase template
Part of story #44077 Artidoc read-only fields: Step results field

To be able to use the tracker in an artidoc, it needs to have the
semantic description configured.

The current template for TestCase have a field named description but not
the semantic

*Testing:*

In a project without TTM, enable the service and start it. The created
TestCase tracker should have the semantic description configured on the
description field

Change-Id: I83abfa5987550bdb15d6140d5255b009f4cfb3da