feat: Add quotes around the service name to delete
Part of request #44088 Service modal should be in typescript

How to test:
No functional change, when you delete a custom service, the translated
message in the confirmation modal should have changed.

Why?
In git #tuleap/stable/2231c641bb0012a7632cb1671d816999b6959d87, I
removed the HTML <b> tags around the service name. However, we can still
wrap it with quotes, it makes it easier for the project admin to
identify the service name that is going to be deleted.

Change-Id: I1f62dd794648994b90daa05c8f8042fe1cae648d