Closes request #44448 README.md file should be interpreted in "Manage Plugin" menu
Since .md is frequently used for README file, it can be a good idea to
support that too. This contribution fix the README.md from the
OpenIdConnect plugin which was not interpreted.

Also, some README.txt file was written in Markdown and was not properly
interpreted, I take the opportunity to convert them to `.md` file

How to test:
 - Go to `Manage Plugins` as Site Admin user
 - See the details of some plugins ( e.g : Open Id Connect, Statistics
   etc...)
=> The README is now correctly displayed
 - Install a plugin.
=> In the modal, the README of the plugin should be also correctly
displayed

Change-Id: Id58aad7ed14317f1df49cceff597b4359735bb39