feat: request #44452 Undo/Redo in prosemirror
prosemirror has a package named history allowing us to have an actions
history in the editor and thus undo/redo actions.

Note: The default delay for actions span is 500ms, I think it is too
much and reduce it by half to 250ms. It can be adjusted easily.

*Testing:*

Write some text in artidoc
- hit Ctrl+z, it should undo what you've typed
- hit Ctrl+y, it should redo what you've typed

Change-Id: Iff61bc6c3fb98a17ae4cefa78029ceffacad817e