refactor: replace TLP global by @tuleap libs in Site admin
part of request #29982 Split `tlp` in smaller packages

How to test:
No functional change expected in Site admin pages:
- In the "Pending users" page, the date pickers for Expiration dates
still work.
- In Permission delegation, when you click on the "Add Permissions"
button, the modal still works.
- When you browse a project and click on the "Change History" tab, the
sub-list still clears when you select an Event type. The date pickers to
filter the list on the left still work.
- When you browse a user details, the date picker for Expiry date still
works.

Why?
I took the opportunity to replace array forEach by `for..of` which is
easier to read.
I also switched the package to type=module so configs must use
ES-modules. It's not mandatory, but it will make it easier to switch to
Vite in the future.

Note:
Since the date-picker is now bundled, the calling scripts get bigger by
about 50 KiB.

Change-Id: Ia7fca8fb65ed0763971f2966047fdcc19f9600b4