fix: request #44455 field xref labels should be aligned
In order to have a consistent design between the "edit mode" artifact
link field and the "readonly mode" field in artidoc, the cross
references should have their left side aligned.

The easiest solution to achieve this rendering is to fix the length of
the link types of each row to 110px. It should be enough to display
everything nicely.

A different approach using css grids have been tested, but it is
actually really difficult to autocompute the link type column based on
the larger label of the set. We could use a fixed column size too, but
it would have the same rendering we have with flex + fixed width, if not
worse.

How to test:
--> xref labels are aligned
--> Same in exported pdfs

Change-Id: Ia7695566ce2c12cacee0a6210f472e152e374f85