chore(👋 svn-core): Explicit instance of Backends
Instantiation of Backend based on const or strings makes life harder
because of weak typing. This change ensures that each time a backend is
instantiated, it has the appropriate type.

The management of singleton was kept to limit the amount of changes but
rewritten with modern PHP capabilities to keep strict typing.

Part of: request #29976 Remove SVN core

Change-Id: Ib9ecc18aaf431043e44fd7e646a0d4218526eb5a