tests: Artidoc REST repeatable test for fields
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
Instead of creating the project only once during the big REST tests
set-up, we define it as a template and create a new project based on it
for each execution. This way, ArtidocFieldsTest is repeatable.
We reuse the Artidoc API and Docman API helpers to get the root folder
ID and create an artidoc in it. Then, we reuse the TrackerRESTHelper to
make it easier to parse the Tracker JSON representation and find the
field IDs. Finally, we PUT the artidoc configuration with the fields. In
a next commit, we will assert that when we have an artifact section in
this artidoc and we GET the section, the payload of the fields match the
format we expect for each type of field.
We also introduce a Disposable to temporarily deactivate the approval of
projects by the site admin. We don't want to break other tests that
might rely on this, so by using Disposable, we make sure it is reset.
We moved the TrackerRESTHelper out of "Regressions", it should be reused
by other tests dealing with trackers.

Co-authored-by: Thomas Gorka <thomas.gorka@enalean.com>
Co-authored-by: Joris Masson <joris.masson@enalean.com>

Change-Id: I03c7b3f1a14b08003d7acdcfcf28d96e979c0fad