refactor: REST API helper for docman
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
Many tests need to get the root folder ID for a given project. I focused
on refactoring ArtidocTest to no longer depend on a test result, but
fetch the ID when it needs it. We use an external helper that can be
composed, instead of a subclass. This will help us move closer to being
able to split this big test file.

Co-authored-by: Thomas Gorka <thomas.gorka@enalean.com>
Co-authored-by: Joris Masson <joris.masson@enalean.com>

Change-Id: Ieb3f17ad07c9dc323fc52f600785ab8cf7806860