refactor: remove PFO_Plugin 🗑️
PFO namespace was introduced in order to improve compatibility between
Fusionforge and Tuleap during Coclico project (which aim was to address
interopabilities and data lock-in in forges).

Since fusionforge project seems to be dead (website is not accessible
and wikipedia says that its latest release was in 2018 – seven years
ago), there is no will for us to maintain dead code.

No functional changes

Part of request #44474: Remove PFO stuff

Change-Id: I0d7eeb3f55188d3180996ab607d879bbafa25dc9