tests: Artidoc fields test creates artifact with all types of fields
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
We want to test the shape of the REST payload for each type of supported
read-only field in Artidoc. We first start by creating an artifact with
all kinds of supported fields. We also import it in the Artidoc. In a
later commit, we will test that getting the section returns the right
payload for each type of field. We do this in steps to make the review
easier.

Change-Id: I0ee002ea1183d911a5bd50f875f74a7da0d0c88e