refactor: REST API helper for core projects
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
Many tests need to get the project ID of a project, but only know its
short name (from the XML file).
I focused on refactoring ArtifactsCreationWithWrongWorkflowTest to no
longer rely on a shared project ID inherited from RestBase. Instead, it
gets the ID by itself. It can seem wasteful, but the goal is to prepare
this test to eventually create its own project and be self-contained and
fully repeatable. It "just" needs a tracker with a workflow, so ideally,
we could create a tracker and set its workflow at the beginning of the
test.
I also refactor TrackerRESTHelper to use field short names, which are
more resilient than labels. Short names are unique for a tracker, labels
are not, and we have both available with XML fixtures.

Change-Id: Ibc31618b2ae4a96a381e27c524825925e2a02bec