refactor: API helper to list user groups of a project
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
We will need to know the IDs of the project's user groups for list field
bound to user groups. It turns out there are two other tests that list
user groups.

I extracted a Disposable for site access (open to anonymous, logged in
or restricted users). It's easier to use the Trait when there are
several small tests, but it's easier to use the Disposable when the test
case is a big scenario so… both will live side by side for now.

I also focused on extracting the part related to listing the user groups
of a project to a dedicated file. This part is fully replayable, though
it still depends on the big set up. We can improve it later.

I was forced to move the UserGroupTest in a namespace now, because
ProjectUserGroupsTest uses the same project, and if the order changes,
it breaks because it will find another user group created by that test.
I choose not to focus on that now, but the proper fix is to reorganize
the tests so that they play a real scenario where the order of things is
expected and is not subject to randomly changing.

Change-Id: I926993bd2d07dfd68a5c77908eaa771063efa330