refactor: replace TLP global by @tuleap/tlp-date-picker in Document
part of request #29982 Split `tlp` in smaller packages

No functional change expected in Date fields in Document app:
- In Advanced search, the date picker still opens to let you choose
dates for criteria
- When you create or update a document, the date picker still opens in
the modals when you have Date custom properties, or obsolescence date.

Note:
Since the date-picker is now bundled, it makes the document app grow:
1,546.64 kB -> 1,596.16 kB

Why?
I forgot to remove the TLP mock in the artifact modal in an earlier
commit. It only affects unit tests and was not used anymore.

Change-Id: I0184193ff3c19c9eb8d8146b608528bb1af9218a