tests: Check all kinds of Artidoc field payloads
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
We run assertions on the REST payloads for each type of supported
read-only field in Artidoc. Only steps definition and steps execution
remain, they will be done in other commits, because they require to
configure another tracker for the Artidoc.

I also refactored ArtidocTest a bit to reuse the code to get the
sections of an artidoc. There is now a dedicated class that should help
tests to run assertions and get data from this big array of sections.

Change-Id: I926102d82d3c6d0635aa01e978ce2ef56b53c4f1