tests: Configure and check payload for Steps definition
part of story #44077 Artidoc read-only fields: Step results field

No functional change. CI should be happy.

Why?
We create a new artidoc to avoid conflicts with the previous step of the
test. This way, we have a brand new artidoc with only one section, that
is configured with the Test Case tracker and with only the Steps
definition field. It could be split to another test file, but I think it
makes sense to group the tests on fields together.

Change-Id: I8be605cabb1c15dca323688a99dfa453afd299ce