feat: Display details per project in a modal
A "details" button is added, which opens a modal to see times per
project for a given user.

Notes:
TLP is updated so that table footer has same padding than header and
body when the table is in a pane section.

The property `label_without_icon` is added to
MinimalProjectRepresentation so that we can sort projects without having
to deal with the icon.

Part of: story #41086: Display times per projects for a given user

Change-Id: Ibb93449e5447d2716eb144fb8a2649efc5203eac