tests: Configure and check payload for Steps execution field
part of request #44829 REST test for Artidoc steps execution field

No functional change. CI should be happy.

Why?
We create another new artidoc to avoid conflicts with the steps
definition part of the test. This way, we have a brand new artidoc with
only one section, that is configured with the Test Execution tracker and
with only the Steps execution field. It could be split to another test
file, but I think it makes sense to group the tests on fields together.

We also need to import a testmanagement XML because we cannot configure
TTM through the REST API.

Change-Id: Ica8ee61ada34617f39066333b5f952a6518a696d