chore(👋 svn-core): move BackendSVN to plugin
Baseline was updated because BackendSVN should be dismantled instead of
being fixed.

The SystemEvent related to project activation is removed as it was there
only for SVN operations but no operations are needed at project
activation related to SVN. It was needed with SVNCore because the
repository as attached to the project itself so at project activation we
had to create the repo. With SVN Plugin, the operations are done when
the repository is created.

Part of: request #29976 Remove SVN core

Change-Id: I2a4944fcf6a19fe6032dfc094c02f9aca985b9ba