fix: PHP implementation of offset/limit/count
The gain in development are modest (ranging from -7% to -11% depending
on the amount of data) but expected to be bigger in production with
large number of artifacts.

Profiling reported that `countArtifactsIdsMatchingQuery` was taking as
much time as `searchArtifactsIdsMatchingQuery`. However, running search
without offset/limit should take the same amount of time. Doing slice
cut and count in php is not expensive so the gain should be significant.

There is a question about memory, if the number of matched artifacts is
really big it could blow-up with this approach.

Closes request #44491 Improve XTS performances on counting

Change-Id: Idef2f7100624950ff68e60412cc689686666e1b7