fix: request #44824 Export project banner in CDATA
Even if project banner is displayed in only one line, the editor allow
to write multiline content. With the CKEDITOR editor it insert non
breaking spaces in the content. While importing the &nbspc; character
break the XML

By wrapping the content in CDATA it fixes the behavior.

*Testing:*

In a project with a multiline banner, export the project structure

- The banner content should be wrapped in a CDATA

You can create a project from the archive

Change-Id: I9d39b90326c0c1260df47ac9d95f3780afeae6f1