feat: introduce feature flag for versions 🚩
To activate display of versions, platform administrators can issue the
following command.
tuleap config-set feature_flag_can_user_display_versions 1

Site administrators can go to site admin » Artidoc, and toggle the
switch.

Please note that implementations of display is not done in this contribution to
keep it small. It will be done in next contribution.

Additional note: an example was added in the past in the design system
to show how to display a TLP switch with a label on the right. Instead
of giving an example and forcing people to copy paste SCSS rules, now it
is part of TLP.

Part of story #44468: Introduce feature flag

Change-Id: I5babc85b8c9d76c8d8798309e8dbbab5b2dcf957