fix: request #44826 Move artifact should ignore permissions when moving
When story #32280 was completed, update permission of fields was
verified and all fields without it were ignored.

As the user doing the move if admin of both source and destination
Trakers, it has all permission on fields.

*Testing:*

Tracker A and B have both 2 same fields

In Tracker B, remove update permission of one field

You should be able to move artifacts from A to B with all the data

Change-Id: I6a64019acb7853e7727becb7ee93494e3caec775