feat: User can choose to display versions
As soon as the current user is a writer of the artidoc document, they
can open the configuration modal to activate the feature.

This is conditionned by the activation of the feature flag in site
admin.

For now it only displays a disabled tab, the fake list of version will
be added later.

You may notice that there is an alignment issue between the tabs and the
toolbar, it will be handled later.

The modal is not suited to display nicely more than two tabs, so it is
switched to medium-sized if user can display version. It avoids the
wrapping of tabs in French locale. (Took the opportunity to remove text
wrapping in fields » display column

Part of story #44468: Introduce feature flag

Change-Id: I9109998bdd6e85b439e193a6280f3b79928d186d