feat: Small fixes on Artidoc site admin settings
part of story #44468: Introduce feature flag

How to test?
Site administrators can go to site admin » Artidoc, and toggle the
switch.

Why?
It feels safer to only allow '1' as a valid value in the form.
The rest are quibbles I did not get a chance to comment on in the
earlier review.

Change-Id: Ifd7a92dd9f43209ee57c8a141d78478aac2e6614