fix: request #44833 Transition rule permissions should apply to project administrators
What has been fixed?
A project admin was able to perform transitions in kanban reports, even is he was not in the group authorized to perform transitions.
This has been fixed and the transition rules now apply to project admin too.

How to test:
-- Having an admin user who is the admin of a project
-- Create an empty group
-- Configure a transition rule for a tracker field and set only this empty group
-- Create a Card wall report with columns based on the same tracker field on which you added the transition rule
-- In this report, drag and drop an artifact into the column where you have set the transition rule

=> A modal should open with the following message:
`The transition to the value "..." is not valid. Transition permissions are not respected`

Change-Id: Ie95647f24479a2cb931a11dd83c6714bdcf69f30