refactor: REST API Helper for TTM
part of request #44829 REST test for Artidoc steps execution field

No functional change. CI should be happy.

Why?
We need to know the IDs of the steps in order to change their status in
the test execution. With a step execution with real step status, we can
test the read-only fields for steps execution field in Artidoc.

I also fixed Psalm issues for TestManagementRESTTestCase and the
associated classes, and for DefinitionsTest.

Change-Id: I998829743e3c2e9394d5a5863863f14af46b7dbf