refactor: Remove date picker JS from TLP
How to test:
- No functional change expected in apps that use TLP date picker. They
all import it from @tuleap/tlp-date-picker, so it should keep working.
For example in Baseline, Document.
- In TLP doc, the date picker is localized in English. The docs explain
how to give it the user locale.

For Site admin pages, if Site admin account has French language, the
date picker should be localized in French:
- In the "Pending users" page, the date pickers for Expiration dates
still work.
- When you browse a project and click on the "Change History" tab, the
date pickers to filter the list on the left still work.
- When you browse a user details, the date picker for Expiry date still
works.
- In Site admin > Global configuration > Platform Banner, the date
picker still works.

For all those pages, the date picker loads its styles.

Why?
The CSS styles are still included in TLP. They will be removed later,
when they are correctly loaded everywhere.
The "datePicker" function is marked as deprecated, the goal is to
replace all its calls gradually everywhere it is used, and to load the
styles at the same time. It defaults the locale to English for now. Once
all the calls are replaced, the date picker will correctly be localized.

part of request #29982 Split `tlp` in smaller packages

Change-Id: Ifcc386945c95e6dccf426f5ca46df1da53ea4a37