refactor: TypeScript target ES2023
part of request #44832 typescript: 5.3.3 -> 5.9.2, tslib: 2.6.3 -> 2.8.1

No functional change. CI should be happy.

Why?
Since the update to TypeScript 5.9, we can now properly target ES2023
(it was not available with TS 5.3). I'm not sure of the impact in
combination with esbuild, but I guess it stops polyfilling some
functions that are now considered defined.

Change-Id: Ifdbf33ba71f82dc3870c853adbeaeb804f11984f