refactor: fix babel-jest type issue
part of request #44832 typescript: 5.3.3 -> 5.9.2, tslib: 2.6.3 -> 2.8.1

No functional change. CI should be happy.

Why?
I had disabled "declaration" when building build-system-configurator,
but it turns out it is useful for vite.config.ts files. Typechecking is
not enforced on them, but our IDE reports an error on viteDtsPlugin. It
should now be fixed.

Change-Id: I56db987c976796fbc1ef118ffd078a118900912a