tests: Graphs snapshot test should not depend on storybook
Implements request #44838 Graphs snapshot test should not depend on
Storybook

No functional change. CI should be happy. When you run `pnpm run test`,
it should not hang forever. Previously, it was stuck forever because in
tlp-doc, vitest would run the "watch" mode instead of the single run
mode.

Why? Vitest in tlp-doc is not configured by the
build-system-configurator, which is why it uses default options and runs
"watch" mode. Since the vite configuration is built by
build-system-configurator in the graphs plugin, it does not have this
problem.

The test does not need to depend upon Storybook at all, so we "just"
call the function on an empty document with the same data options passed
by Storybook, and it runs correctly.

Change-Id: I3df675ffa5173aa12f0f38fc77eaecd5058eb821