feat: User can switch between TOC and Versions
When the user activates the display of versions (will need the feature
flag as site admin for that), then the Versions tab is not anymore
disabled and user can see its content.

For now the content is useless, it will be implemented later. This
contribution focuses on the tabs mechanism (same as configuration
modal).

Closes story #44468: Introduce feature flag

Change-Id: Ia937044667cec1721b997bda1bb63d1166e66a8c