tests: Fix flakiness of artifact_link-usage, cover legacy links field
part of request #44842 e2e test for artifact link types can be flaky

How to test:
No functional change. The following command should pass:
make tests-e2e

Why?
The flakiness stemmed from the fact that Cypress could sometimes find
two link type selects. Waiting for the links to be loaded (by checking
the skeletons are gone) and scoping the selection better should stop it
from being flaky.

We also now reuse our cy.searchItemInLazyboxDropdown() command to handle
adding new links and creating a new artifact.

Assertions on Parent/Child have been remade to check that the Title
shows the parent artifact.

The test for creating a new artifact has been moved to the "artifact"
project context, so that all tests using the new artifact links are
together. I added the Title semantic on the tracker (and dropped unused
services from the project) to be able to create a new artifact (title is
required).

It turns out the "can use _is_child type" test did almost the exact same
thing as the "can create a new Parent link" and "can update a Parent
link" tests. Instead of deleting tests, I realized that the legacy
artifact links field was no longer covered by e2e tests. That is why I
chose to transform them back to uncheck the specific property and test
the legacy field. Since there is special magic in the backend to handle
the Parent type with the legacy field, it makes sense to cover it with a
dedicated test for "can create a new Parent link".

I also took the opportunity to remove some dynamic data-test attributes.
data-test should be static, we should rely on the text the user can see
and build our selections the same way a user would interpret it: find
the line with the link type we are looking for, and in this line, find
the right button. This is why the Site admin and Tracker admin tests
have also changed.

Change-Id: I3f4b7412175b75dd22aed4c20631b2e15ca6b942