feat: request #44822 Make possible to force users to update password after a certain date
A new setting sys_password_expiration_date is added. When both
sys_password_expiration_date and sys_password_lifetime are defined the
password is considered expired as soon as one of the limit is reached.
When only one of these restrictions are defined, the password is
considered expired as soon as the limit is reached.

The Clock abstraction usage is made explicit, some parts of Tuleap were
already relying on it anyway.

Change-Id: I0e15fd23d25e90f9696da80af04c4a96980bdae3