feat: Artifact view dates are displayed with submitter timezone
Part of request #44843 Date fields should be displayed with submitter timezone

For date field we show a day, and a day is a day and the same day for
everyone no matter your timezone.

Currently depending on your timezone you may not see the same day than
other users on another timezone: we store day with the time set at
00:00. So if you are on UTC and your friend on UTC-4, for your friend
the date is the previous day at 20:00 but displayed without the time
(confusing)

To fix that, for date fields (without time) we display the date with the
timezone of the changeset submitter in artifact view.

Note: this is hidden behind `display_date_with_submitter_timezone`
config key and available only in artifact view

*Testing:*

Enable the feature with `tuleap config-set display_date_with_submitter_timezone 1`

In a tracker with a date field (no time), create an artifact with a
given date in a certain timezone (Europe/Paris for example).

With another user with a different timezone (America/Los_Angeles for
example), fetch this artifact -> the displayed date should be the same
day.

Change-Id: I8b48d0525bc09d8c8e1e2855c2044d385fa4324b