refactor: Use LinkDirection more
part of request #44820 Parent artifacts should be displayed first in
readonly links field

No functional change. CI should be happy.

Why?
In git #tuleap/stable/1a4c45edeb3bbd10aa411a773789dc6c3f0b7697, we
introduced the LinkDirection enum to better represent the direction
"forward" or "reverse". We should try to use it as much as possible,
while keeping things safe.

I also tried to reuse the constants in ArtifactLinksField more, so that
we have a chance of identifying all the cases where we need '' to
represent "no type".

I chose to allow using LinkDirection in Artidoc Domain because it's
"just" an enum. In my opinion, there is more value in reusing it instead
of copying it to preserve the Domain purity or falling back to "string".

Change-Id: Ib25476b233d591eff6c642550c8bfd189cc9b785