fix: request #44852 Wrong arrow positionning after expanding a row in XTS
When the `@pretty_title` is not the fist column of the query, if there are already
drawn arrows and a new row is expanded, old arrows will be "pushed to the right" of the table.

This bug is due to a nice mix between the skeletons inaccurate and
the ResizeObserver not catching any resizing. First, skeletons are
inaccurate in terms of wideness of some columns: for instance the
skeleton for the `@id` column is much wider than the resulting column.
This has the direct effect of pushing the next columns (@pretty_title)
for instance to the right on load. Because skeletons are introduced in
the table on loading, the table size changes and a redraw of the arrows
is triggered, with the columns effectively pushed to the right and thus
arrows drawn there. Once the data is loaded, skeletons are swapped with
the real content. Because the height of skeletons is accurate, when
swapping them with their data counterpart there is no size change of the
table and thus the ResizeObserver does not catch the change and no
resize event is raised so no arrows are redrawn: the old arrows stay on
the right.

*How to test:*

1. Have a query where `@pretty_title` is not in the first position, for
   instance:

   ```
    SELECT @id, @pretty_title, @tracker.name, @project.name, @last_update_date, @submitted_by
    FROM @tracker.name IN ('story')
    WHERE @id > 1
   ```

2. Expand a few rows
    a. All arrows should be correctly placed instead of some being pushed
    to the right

Change-Id: Id8ab417f521f683c15bdeb26806aa39bed111614