feat: Display fake versions description
By default, only the first words are displayed with an ellipsis at the
end.
Description can be collapsed/expanded to have the full text.

If the description is small enough, then expand/collapse button is not
displayed. This depends on the size of the viewport, so we need to
observe it.

Part of story #44469: Visualize list of versions

Co-authored-by: Nicolas Terray <nicolas.terray@enalean.com>
Co-authored-by: Joris Masson <joris.masson@enalean.com>

Change-Id: Ifba2ad51014cb9e08951541f6bab2564f0e50753