refactor(cross-tracker-search): Order the row collection and remove listeners of TableDataStore
part of story #44821 Include linked artifacts to XLSX export

This change modifies the behavior of the TableDataStore. First,
it removes the listeners from TableDataStore now relying on the
provide/inject mechanism to make it available to all concerned
components. Second, the row collection is now ordered as they would be
shown in the result table, meaning that children are below their parents
and reverse links are after the forward ones.

*No functional changes expected*

Co-authored-by: Marie-Ange Garnier <marie-ange.garnier@enalean.com>
Change-Id: If15e574604a8032913083d6fa14e24825844afb7