refactor: Load date picker only for site admin User registration
How to test:
- When not logged in, visit /account/register.php (or click on "New
User") button in the top bar).
- No functional change expected.
- As site admin, visit /admin/register (or go to Users > "New user" in
the sidebar)
- The "Expiration date" date picker should still work. If Site admin
account has French language, the date picker should be localized in
French. It loads the date picker styles.
- No other functional change: the "Generate password" button still
works.

Why?
In previous changes to replace usage of global "tlp" by
@tuleap/tlp-date-picker, we forgot to localize the date picker correctly
and to load its CSS styles. The goal is to completely remove the CSS
from TLP.

We don't need to load KBs of date picker for anonymous user who will not
even use them (the date picker HTML is only added for site admin). Thus,
we introduce a second endpoint that is dedicated for site admin.

part of request #29982 Split `tlp` in smaller packages

Change-Id: Iacc047bfd99a5e392afc6c417b340f23fee9f7e0