feat: add a11y button to display a section
Someone using a keyboard should be able to display a given version.

Hit tab until you reach a version in the sidebar and press enter or
space.

Since the target audience of this button is people using keyboard or
screen readers, the button is not display when hovering the section with
the cursor in order to not clutter the interface.

Part of story #44469: Visualize list of versions

Change-Id: Ie8af14ade72ea5d944743a44e55c6418b2b06e2b