fix: alignement issue in condensed mode
When a user is in condensed mode, then there is an alignment issue
between the sidebar header and the toolbar. This is due to the fact that
the min-height of the tlp-tabs is defined on --tlp-jumbo-spacing, which
depends on condensed mode.

There should be no alignment issue on both Condensed and Comfortable
modes.

Part of story #44469: Visualize list of versions

Change-Id: If0f218565b0f087571e3f5af7964348f9ff9777d