refactor: use Tuleap locale string format in Document
part of request #29982 Split `tlp` in smaller packages

How to test:
No functional change in Document:
- Relative dates are still translated to your user language.
- You can still create new ONLYOFFICE docs according to your language.

No functional change in Roadmap:
- the months are still localized when selecting "Months" time period
- the "today" indicator still has a title with the translated date.

Why?
The "toBCP47" function is now moved to @tuleap/locale to be reused.
There is still a copy left in the artifact modal, but it will require
more work and a bigger patch, we should de-dupe it later.

I chose to use the basic "LocaleString" type for the user locale in
Document, and to switch it to BCP47 format only where needed (for
example, for tlp-relative-date elements). The tlp date picker expects a
LocaleString, and in the next commit we will be able to inject it
directly.

Change-Id: I5e3ddcc2526d8325c96877b0f8ffb962a5e4f23e