refactor: Extract loadProjectProperties from properties-actions
Part of request #44447 Get rid of vuex in document

This was the last actions in properties. As it was also acting on the
only mutations to load the state of properties. Extracting it resulted
in removing everything, hence all the modifications.

As the state for project_properties no longer exists, I choose to
strictInject a ref, this way components who call loadProjectProperties
set the value of the ref, and components who get project_properties have
the update with the ref

*No functional changes expected*

Change-Id: I8dd8ddb2131e56ccb49123813edc07bdf7b6d0e3