fix: request #45227 no change submission removes the artifact link field
What has been fixed?
When you click on a field to edit it and then submit without changing anything, it puts all fields into edit mode and the link field disappears.
This behavior has been changed so that after submission, you are redirected to the tracker page (or the artifact page if you clicked on submit and stay).
This improves the user experience by matching the behavior of a normal submission.
Note that a feedback informs the user that no changes have been made.
This also fixes the problem of the links field disappearing.

How to test:
-- On Artifact view, click on a field to edit it
-- Click on Submit and stay button without changing anything
=> It should display the Artifact view with a feedback indicating that nothing have been changed

Change-Id: I5a431d373a120b514fc4d9bd2478ecc7aa066e8a