chore: Scripts in PHPWiki templates have a nonce
Login templates are removed, it is not something that can be reached
within Tuleap context.
Redirect templates are also not useful, the redirection are done using
HTTP headers.
The editpage templates are not meaningful for the themes used in the
printer versions.

Part of request #17967: Deploy a useful content security policy

Change-Id: Ie35ab7e534d4cbdfd725fc0f6a4b51971947340c