refactor: localize date pickers in Personal timetracking widget
part of request #29982 Split `tlp` in smaller packages

How to test:
- If you current user account has French language, the date pickers in
the Personal Timetracking widget should be localized in French.
- The date picker styles should be loaded.
- No other functional change expected.

Why?
In previous changes to replace usage of global "tlp" by
@tuleap/tlp-date-picker, we forgot to localize the date picker correctly
and to load its CSS styles. The goal is to completely remove the CSS
from TLP.

Also, we switch to provide/inject instead of pinia store for user
locale, in accordance with ADR-0033.

Change-Id: I18faa5c64e93d5312fb340486aba2815768293b3