fix: request #45224 Searching artifacts via cross references can be slow
The existing indexes can not be used due to a cast issue.
We now force the cast in the same way we do for the permissions table.

This is outside the scope of this contribution but at some points we
will need to adjust the deprecated utf8 character set across the whole
codebase.

Change-Id: I0e25a23d76ed13c878afaba78a843564419555c3