refactor: Extract embedded display preference from preferencies-actions
Part of request #44447 Get rid of vuex in document

Have started only with getEmbeddedFileDisplayPreference but have to also
the 2 other actions, the mutation and the state as they all overlap with
the others.

Now when displaying a embedded document, the top component
DisplayEmbedded retrieve the user preference and pass it as a props to
the subcomponent. EmbeddedFileEditionSwitcher call the api to persist
the change and emit a vue event up to DisplayEmbedded to change the ref
and propagate the change to child components

There is still setUserPreferenciesForFolder to extract

*No functional changes expected*

You can still change your display preference per document, and it is
persisted in time

Change-Id: If6af6d50cde868f99260669d8f3cbf6d143bdf15