chore: fa-solid for save, asterisk, users, plus
Prior to FontAwesome 6, the usage of icon was done like this: `<i
class="fa fa-save"></i>` Starting FontAwesome 6, the classname `fa` is
converted to `fa-solid` or `fa-regular` in order to choose a style for
the icon.

While the recommendation is to use FontAwesome 6, there are still parts
of the code that use legacy `fa fa-save`. And since developers copy
paste snippets of code for new contributions, in the code review we are
sometimes rejected because we didn't see that we need to convert to
`fa-solid fa-save`.

Part of request #45229: Use FontAwesome 6 classname for style

Change-Id: Ic8ec351f3341db879e5e76f236483f52be8c1617