chore: fa-solid for circle-{info,exclamation} triangle-exclamation
Prior to FontAwesome 6, the usage of icon was done like this: `<i
class="fa fa-save"></i>` Starting FontAwesome 6, the classname `fa` is
converted to `fa-solid` or `fa-regular` in order to choose a style for
the icon.

While the recommendation is to use FontAwesome 6, there are still parts
of the code that use legacy `fa fa-save`. And since developers copy
paste snippets of code for new contributions, in the code review we are
sometimes rejected because we didn't see that we need to convert to
`fa-solid fa-save`.

Part of request #45229: Use FontAwesome 6 classname for style

Change-Id: Ieefcdbba7b9c2852f507aa5f68b3d279da1ac787