feat(ai): Introduce Cross-Tracker Search Assistant plugin
This is the scaffolding of the plugin. The goal is to introduce a base
route that highlight how we can exchange with Mistral API.

The route itself is mainly there to test interaction with the API and
find the final design of the system.

Part of: story #44884 TQL assistant prototype

Change-Id: I13bd5d2adbb8757a87d1da0002c5a2e4dc9ca496